additional refactoring to streamline the package
