+++ token/core/zkatdlog/nogh/v1/audit/testdata/idemix/msp/RevocationPublicKey
