+++ token/core/zkatdlog/nogh/v1/setup/setup.go
@@ -4,7 +4,7 @@ Copyright IBM Corp. All Rights Reserved.
 SPDX-License-Identifier: Apache-2.0
 */
 
-package crypto
+package setup
 
 import (
 	"crypto/sha256"
