+++ token/core/zkatdlog/nogh/v1/setup/setup_test.go
@@ -3,7 +3,7 @@ Copyright IBM Corp. All Rights Reserved.
 
 SPDX-License-Identifier: Apache-2.0
 */
-package crypto
+package setup
 
 import (
 	"fmt"
