+++ token/core/zkatdlog/nogh/v1/setup/testdata/idemix/msp/RevocationPublicKey
