--- token/core/zkatdlog/nogh/v1/validator/ecdsa/des.go
@@ -1,40 +1,0 @@
-/*
-Copyright IBM Corp. All Rights Reserved.
-
-SPDX-License-Identifier: Apache-2.0
-*/
-
-package ecdsa
-
-import (
-	"bytes"
-
-	"github.com/hyperledger-labs/fabric-smart-client/pkg/utils/errors"
-	"github.com/hyperledger-labs/fabric-token-sdk/token/driver"
-	"github.com/hyperledger-labs/fabric-token-sdk/token/services/identity"
-	"github.com/hyperledger-labs/fabric-token-sdk/token/services/identity/x509"
-	"github.com/hyperledger-labs/fabric-token-sdk/token/services/utils"
-)
-
-type Deserializer struct {
-	auditInfo []byte
-}
-
-func (d *Deserializer) Match(id []byte) error {
-	identity, err := identity.WrapWithType(x509.IdentityType, id)
-	if err != nil {
-		return errors.Wrapf(err, "failed to unmarshal identity [%s]", id)
-	}
-	if !bytes.Equal(d.auditInfo, identity) {
-		return errors.Errorf("identity mismatch [%s][%s]", utils.Hashable(identity), utils.Hashable(d.auditInfo))
-	}
-	return nil
-}
-
-func (d *Deserializer) GetAuditInfoMatcher(owner driver.Identity, auditInfo []byte) (driver.Matcher, error) {
-	return &Deserializer{auditInfo: auditInfo}, nil
-}
-
-func (d *Deserializer) DeserializeVerifier(id driver.Identity) (driver.Verifier, error) {
-	panic("implement me")
-}
